Refactored register widget for reuse
