+++ lib/pages/auth/login.dart
@@ -14,7 +14,7 @@ class AuthLoginPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AuthMainWidget(
-      child: AuthLoginWidget(email: email),
+      child: AuthLoginWidget(email),
       header: "Sign in to your account",
       subHeader: "Enter your password to continue.",
     );
