--- lib/pages/auth/sign_in.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../widgets/index.dart';
-
-class AuthSignInPage extends StatelessWidget {
-  const AuthSignInPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const AuthMainWidget(
-      child: AuthSignInPageWidget(),
-      header: "Sign in to your account",
-      subHeader: "Enter your password to continue.",
-    );
-  }
-}
