+++ lib/widgets/auth/email_password/mixin.dart
@@ -1,0 +1,5 @@
+import '../../../router/index.dart';
+
+mixin AuthEmailPasswordMixin {
+  String? onPasswordSubmit(RouteProvider router, String password);
+}
