+++ lib/widgets/auth/login.dart
@@ -1,16 +1,25 @@
 import 'package:flutter/material.dart';
 
-class AuthLoginWidget extends StatelessWidget {
+import '../../router/index.dart';
+import 'email_password/email_password.dart';
+import 'email_password/mixin.dart';
+
+class AuthLoginWidget extends StatelessWidget with AuthEmailPasswordMixin {
   final String email;
 
-  const AuthLoginWidget({
-    required this.email,
-    Key? key,
-  }) : super(key: key);
+  const AuthLoginWidget(this.email, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    // TODO: Implement login
-    return Container();
+    return AuthEmailPasswordWidget(
+      email: email,
+      onPasswordSubmit: onPasswordSubmit,
+    );
+  }
+
+  @override
+  String? onPasswordSubmit(RouteProvider router, String password) {
+    // TODO: onPasswordSubmit
+    return null;
   }
 }
