--- lib/widgets/auth/page/sign_in.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AuthSignInPageWidget extends StatelessWidget {
-  const AuthSignInPageWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    // TODO: register.dart
-    return Container();
-  }
-}
